# Optimize parsing of trace-cmd output