The combined effect of these optimizations is it reduced the time take to loop
over all events in a trace containing 1.5M events from ~70s to ~11s.

- Not using regex to parse events saves ~39s
- Not parsing event bodies unnecessarily saves ~18s*
- Not matching every line against preamble/dropped events regex's saves ~2s

*The benchmark just looped over events; it did not try to access any fields,
 so no bodies were parsed. A typical script would usually access fields in
 some of the events, so this optimization would not be as impactful, but it
 should still effect a significant saving by not parsing all of event bodies.